requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
reportlab==4.4.4
six==1.17.0
sqlparse==0.5.3
gunicorn
whitenoise

